removed HEAD support while caching index for remote Forban
The HEAD was introduced to permit the comparison between a cached
index and a remote index. The comparison was done only on the size
of the file. But this is a bad idea, as the size of a index could
give the impression that the index didn't change but some values
where inversed (especially the size value after each line). This
was introducing a bug where a large download is taking place and
files were growing in size. So the fetching of the files was based
on the wrong data in the index...

A new scheme for checking index change need to be done (especially
if the indexes are quite large among the Forban). Maybe the
HMAC value of the announce could be used to announce an updated
index...