fixed #6 (symbolic links followed for index building)
Index building is now following symbolic links. As this was requested
for low-end hardware where the shared files are shared among flash, ram-based
fs... There are security risks to include symbolic links that should
be considered when adding links in the shared directory.

CherryPy is already supporting symbolic links for the HTTP server.